+++ src/estd/algorithm.h
@@ -51,7 +51,7 @@ inline ESTD_CPP_CONSTEXPR(17) void fill(ForwardIt first, ForwardIt last, const T
 
 // https://en.cppreference.com/w/cpp/algorithm/equal
 template<class InputIt1, class InputIt2>
-bool equal(InputIt1 first1, InputIt1 last1,
+ESTD_CPP_CONSTEXPR(17) bool equal(InputIt1 first1, InputIt1 last1,
            InputIt2 first2)
 {
     for (; first1 != last1; ++first1, ++first2) {