+++ src/estd/internal/array.h
@@ -182,4 +182,20 @@ struct allocator_traits<internal::layer1_allocator<T, N> >
 
 };
 
+template <class Impl1, class Impl2>
+constexpr bool operator==(
+    const internal::array<Impl1>& lhs,
+    const internal::array<Impl2>& rhs)
+{
+    return lhs.size() == rhs.size() ? equal(lhs.begin(), lhs.end(), rhs.begin()) : false;
+}
+
+template <class Impl1, class Impl2>
+constexpr bool operator!=(
+    const internal::array<Impl1>& lhs,
+    const internal::array<Impl2>& rhs)
+{
+    return lhs.size() == rhs.size() ? !equal(lhs.begin(), lhs.end(), rhs.begin()) : true;
+}
+
 }